+++ app/src/main/java/com/smaher/tasktimerapp_pinkninjas/EditFragment.kt
@@ -87,7 +87,7 @@ class EditFragment : Fragment() {
             if(title.text.isNotBlank()
                 &&description.text.isNotBlank()
                 &&time.text.isNotBlank()){
-                myViewModel.updateTask(Task(0,title.text.toString(),description.text.toString(),type,"new",time.text.toString().toInt()))
+                myViewModel.updateTask(Task(0,title.text.toString(),description.text.toString(),type,"new",time.text.toString().toLong()))
                 Toast.makeText(this.context,"Task Added successfully",Toast.LENGTH_SHORT).show()
                 hideKeyboard()
                 Navigation.findNavController(view).navigate(R.id.action_editFragment_to_homeFragment)
